+++ TKPEmu/emulator.h
@@ -29,7 +29,7 @@ namespace TKPEmu {
 		virtual void Reset() = 0;
 		virtual void Update() = 0;
 		virtual void LoadFromFile(std::string&& path) = 0;
-		virtual uint8_t* GetScreenData() { return nullptr; };
+		virtual float* GetScreenData() { return nullptr; };
 		std::mutex ThreadStartedMutex;
 		std::mutex DrawMutex;
 		std::thread UpdateThread;
